refactor: centralize UI helpers
