+++ src/step2.js
@@ -10,6 +10,7 @@ import {
   MAX_CHARACTER_LEVEL,
 } from './data.js';
 import { t } from './i18n.js';
+import { createElement, createAccordionItem } from './ui-helpers.js';
 
 const abilityMap = {
   Strength: 'str',
@@ -100,46 +101,6 @@ function validateTotalLevel(pendingClass) {
     return false;
   }
   return true;
-}
-
-function createElement(tag, text) {
-  const el = document.createElement(tag);
-  if (text) el.textContent = text;
-  return el;
-}
-
-function createAccordionItem(title, content, isChoice = false, description = '') {
-  const item = document.createElement('div');
-  item.className = 'accordion-item' + (isChoice ? ' user-choice' : '');
-
-  const header = document.createElement('button');
-  header.className = 'accordion-header';
-  if (description) {
-    const titleSpan = document.createElement('span');
-    titleSpan.textContent = title;
-    const descSpan = document.createElement('small');
-    descSpan.textContent = ` - ${description}`;
-    header.appendChild(titleSpan);
-    header.appendChild(descSpan);
-  } else {
-    header.textContent = title;
-  }
-
-  const body = document.createElement('div');
-  body.className = 'accordion-content';
-  if (typeof content === 'string') {
-    body.textContent = content;
-  } else {
-    body.appendChild(content);
-  }
-  header.addEventListener('click', () => {
-    header.classList.toggle('active');
-    body.classList.toggle('show');
-  });
-
-  item.appendChild(header);
-  item.appendChild(body);
-  return item;
 }
 
 function getProficiencyList(type) {
